services/clip-embedding-service/requirements.txt
fastapi
uvicorn
sentence-transformers
Pillow
torch
torchvision
